plus de nasm

+++ nasm/make.bash
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-nasm -f elf64 "$1"
+if [ ! -f "libIUT/libIUT.a" ]; then
+  cd libIUT || exit
+  make libIUT.a
+  cd ..
+fi
+
+nasm -g -f elf32 "$1"
 name="$(basename "$1" .nasm)"
-gcc -o "$name" "$name".o -L. -lIUT -lncurses
+gcc -m32 -o "$name".out "$name".o -LlibIUT -lIUT -lpanel -lncurses
 rm "$name".o
